Related work items: #7944, #8223, #8224

+++ Code/Avalanche.Shared.Domain/Models/Region.cs
@@ -6,9 +6,9 @@ namespace Avalanche.Shared.Domain.Models
 {
     public class Region
     {
-        public int Height { get; set; }
-        public int Width { get; set; }
-        public int X { get; set; }
-        public int Y { get; set; }
+        public decimal Height { get; set; }
+        public decimal Width { get; set; }
+        public decimal X { get; set; }
+        public decimal Y { get; set; }
     }
 }
